appstore: small refactoring, proper cache/refresh mechanism for catalog and wishlist
