+++ store/config.toml.example
@@ -7,4 +7,3 @@ GITHUB_LOGIN = "yunohost-bot"
 GITHUB_EMAIL = "yunohost [at] yunohost.org" # Replace the [at] by actual @
 GITHUB_TOKEN = "superSecretToken"
 APPS_CACHE = "../.apps_cache/"
-STARS_DB_FOLDER = ".stars/"
